add multi-ask test
